# Replace standard json library by segmentio's library