--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -10,14 +10,11 @@ jobs:
     strategy:
       matrix:
         go:
+          - "1.21"
           - "1.20"
           - "1.19"
           - "1.18"
           - "1.17"
-          - "1.16"
-          - "1.15"
-          - "1.14"
-          - "1.13"
 
     steps:
       - name: Set up Go ${{ matrix.go }}