--- a/cfn/event_test.go
+++ b/cfn/event_test.go
@@ -3,7 +3,7 @@
 package cfn
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"io/ioutil" //nolint: staticcheck
 	"testing"