--- a/cfn/wrap_test.go
+++ b/cfn/wrap_test.go
@@ -5,8 +5,8 @@ package cfn
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
+	"github.com/segmentio/encoding/json"
 	"io/ioutil" //nolint: staticcheck
 	"net/http"
 	"testing"