--- a/events/README_KinesisDataAnalytics.md
+++ b/events/README_KinesisDataAnalytics.md
@@ -10,7 +10,7 @@ package main
 
 import (
 	"context"
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"fmt"
 	"log"