--- a/events/apigw_test.go
+++ b/events/apigw_test.go
@@ -3,7 +3,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"io/ioutil" //nolint: staticcheck
 	"testing"