--- a/events/attributevalue.go
+++ b/events/attributevalue.go
@@ -5,9 +5,9 @@ package events
 import (
 	"bytes"
 	"encoding/base64"
-	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/segmentio/encoding/json"
 	"strconv"
 )