--- a/events/chime_bot_test.go
+++ b/events/chime_bot_test.go
@@ -2,7 +2,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"testing"
 	"time"