--- a/events/clientvpn_test.go
+++ b/events/clientvpn_test.go
@@ -1,7 +1,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"io/ioutil" //nolint: staticcheck
 	"testing"