--- a/events/cloudwatch_events.go
+++ b/events/cloudwatch_events.go
@@ -1,7 +1,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"time"
 )