--- a/events/cloudwatch_logs_test.go
+++ b/events/cloudwatch_logs_test.go
@@ -1,7 +1,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"reflect"
 	"testing"