--- a/events/connect_test.go
+++ b/events/connect_test.go
@@ -3,7 +3,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"testing"
 
 	"github.com/aws/aws-lambda-go/events/test"