--- a/events/ecr_image_action_test.go
+++ b/events/ecr_image_action_test.go
@@ -2,7 +2,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"testing"
 	"time"