--- a/events/ecs_container_instance.go
+++ b/events/ecs_container_instance.go
@@ -1,7 +1,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"time"
 )