--- a/events/ecs_container_instance_test.go
+++ b/events/ecs_container_instance_test.go
@@ -2,7 +2,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"testing"
 	"time"