--- a/events/epoch_time.go
+++ b/events/epoch_time.go
@@ -3,7 +3,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"time"
 )