--- a/events/epoch_time_test.go
+++ b/events/epoch_time_test.go
@@ -2,7 +2,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"testing"
 	"time"