--- a/events/firehose_test.go
+++ b/events/firehose_test.go
@@ -3,7 +3,7 @@ package events
 
 import (
 	"context"
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"strings"
 	"testing"