--- a/events/iot_preprovision_hook_test.go
+++ b/events/iot_preprovision_hook_test.go
@@ -1,7 +1,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"io/ioutil" //nolint: staticcheck
 	"testing"