--- a/events/kafka.go
+++ b/events/kafka.go
@@ -3,7 +3,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 )
 
 type KafkaEvent struct {