--- a/events/lambda_function_urls_test.go
+++ b/events/lambda_function_urls_test.go
@@ -3,8 +3,8 @@
 package events
 
 import (
-	"encoding/json"
 	"errors"
+	"github.com/segmentio/encoding/json"
 	"io/ioutil" //nolint: staticcheck
 	"net/http"
 	"strings"