--- a/events/s3.go
+++ b/events/s3.go
@@ -3,7 +3,7 @@
 package events
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"net/url"
 	"time"
 )