--- a/events/test/assert.go
+++ b/events/test/assert.go
@@ -1,7 +1,7 @@
 package test
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"io/ioutil" //nolint: staticcheck
 	"testing"