--- a/events/test/jsonsyntax.go
+++ b/events/test/jsonsyntax.go
@@ -3,7 +3,7 @@
 package test
 
 import (
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"testing"
 )