--- a/go.mod
+++ b/go.mod
@@ -1,11 +1,16 @@
 module github.com/aws/aws-lambda-go
 
-go 1.18
+go 1.21
 
-require github.com/stretchr/testify v1.7.2
+require (
+	github.com/segmentio/encoding v0.3.6
+	github.com/stretchr/testify v1.8.4
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/segmentio/asm v1.2.0 // indirect
+	golang.org/x/sys v0.13.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )