--- a/lambda/rpc_function.go
+++ b/lambda/rpc_function.go
@@ -7,8 +7,8 @@ package lambda
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
+	"github.com/segmentio/encoding/json"
 	"log"
 	"net"
 	"net/rpc"