--- a/lambdaurl/testdata/testfunc/site/main.go
+++ b/lambdaurl/testdata/testfunc/site/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"context"
-	"encoding/json"
+	"github.com/segmentio/encoding/json"
 	"io"
 	"net/http"
 	"os"