@bmoffatt 2023-10-30
I'm not comfortable expanding the dependencies of aws-lambda-go beyond that of Go's standard library.